+++ nest/src/authentication/auth.service.ts
@@ -1,58 +1,35 @@
-import {
-  BadRequestException,
-  HttpException,
-  HttpStatus,
-  Injectable,
-  NotFoundException,
-  UnauthorizedException,
-} from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import * as bcrypt from 'bcrypt';
-import { v4 as uuidv4 } from 'uuid';
 import { ConfigService } from '@nestjs/config';
 import { ICookieProps, IJwtPayload, JwtUserProps } from '@typing/auth';
 import { addMinutes, millisecondsToMinutes } from 'date-fns';
 import { IncomingHttpHeaders } from 'http';
 import makeHash from '@common/helpers/make-hash.helper';
 import jwtVerifyPromise from '@common//helpers/jwt-promise.helper';
-import UserVerifyEntity from '@models/user/entities/user-verify.entity';
 import UserVerifyRepository from '@models/user/repositories/user-verify.repository';
 import UserRepository from '@models/user/repositories/user.repository';
-import UserJobRepository from '@models/user/repositories/ user-job.reposity';
 import UserEntity from '@models/user/entities/user.entity';
 import UserService from '@models/user/user.service';
 import CreateUserDto from '@authentication/dto/create-user.dto';
+import AuthValidateService from '@authentication/auth-validate.service';
 
 @Injectable()
-class AuthService {
+export default class AuthService {
   constructor(
     private configService: ConfigService,
     private userService: UserService,
+    private authValidateService: AuthValidateService,
     private jwtService: JwtService,
     private userVerifyRepository: UserVerifyRepository,
     private userRepository: UserRepository,
-    private userJobRepository: UserJobRepository,
   ) {}
-
-  /**
-   * @return 로그인 성공 시 패스워드가 제외된 유저 정보 객체를 넘긴다.
-   */
-  async validateUser(email: string, password: string) {
-    const user = await this.userRepository.findUserByEmailForLogin(email);
-    if (user === null) throw new NotFoundException('가입되지 않은 유저입니다.');
-    const isPasswordMatch = await bcrypt.compare(password, user.password);
-    if (!isPasswordMatch) {
-      throw new UnauthorizedException('비밀번호가 틀렸습니다.');
-    }
-    const { password: notUsingProp1, hashedRefreshToken: notUsingProps2, ...returnProps } = user;
-    return returnProps;
-  }
 
   /**
    * @desc id 에 대응하는 사용자의 refreshToken 을 제거한다.
    */
-  async removeRefreshToken(id: number) {
-    await this.userRepository.update(id, {
+  removeRefreshToken(id: number): void {
+    this.userRepository.update(id, {
       hashedRefreshToken: null,
     });
   }
@@ -60,7 +37,7 @@ class AuthService {
   /**
    * @return AccessToken 을 생성하여 쿠키 정보와 함께 반환한다.
    */
-  getAccessToken({ id, username }: JwtUserProps) {
+  getAccessToken({ id, username }: JwtUserProps): string {
     const accessTokenExpirationTime = Number(
       this.configService.get('JWT_ACCESS_TOKEN_EXPIRATION_TIME'),
     );
@@ -75,7 +52,7 @@ class AuthService {
   /**
    * @return RefreshToken 을 생성하여 객체정보로 반환한다.
    */
-  getRefreshTokenCookie({ id, username }: JwtUserProps) {
+  getRefreshTokenCookie({ id, username }: JwtUserProps): ICookieProps {
     const refreshTokenExpirationTime = Number(
       this.configService.get('JWT_REFRESH_TOKEN_EXPIRATION_TIME'),
     );
@@ -89,143 +66,41 @@ class AuthService {
       token,
       maxAge: 6.048e8,
       path: '/',
+      httpOnly: true,
     } as ICookieProps;
   }
 
   /**
    * @return RefreshToken 검증 후, 사용자 정보 객체를 반환합니다.
    */
-  async getUserIfRefreshTokenMatches(refreshToken: string, id: number) {
-    const user = (await this.userRepository.findOne(id)) as UserEntity;
-
+  async getUserIfRefreshTokenMatches(refreshToken: string, id: number): Promise<UserEntity | null> {
+    const user = await this.userRepository.findOne(id);
+    if (!user) return null;
     const isRefreshTokenMatching = await bcrypt.compare(
       refreshToken,
       user.hashedRefreshToken as string,
     );
-
-    if (isRefreshTokenMatching) {
-      return user;
-    }
+    if (isRefreshTokenMatching) return user;
     return null;
-  }
-
-  /**
-   * @return 이메일 형태를 검증합니다.
-   */
-  verifyEmailRequest(email: string | undefined) {
-    if (!email) {
-      throw new HttpException('이메일 필드가 존재하지 않습니다.', HttpStatus.BAD_REQUEST);
-    }
-
-    const isEmpty = email.trim() === '';
-    const matcher = email.match(/\w+@\w+.\w{3}/);
-    if (isEmpty || matcher === null) {
-      throw new HttpException('이메일 형식이 아닙니다.', HttpStatus.BAD_REQUEST);
-    }
-  }
-
-  /**
-   * @desc 이메일을 입력받습니다.
-   * 이메일을 통해 이미 레코드가 있는지 보고, 레코드의 만료 날짜가 지났는지 검증합니다.
-   * 만료 날짜가 아직 지나지 않았다면 기존에 생성된 레코드를 그대로 리턴할 것입니다.
-   * 그렇지 않다면 새로운 레코드를 만들어서 프론트에 제공합니다.
-   */
-  async getEmailVerifyToken(email: string) {
-    const now = new Date();
-    const currentVerification = await this.userVerifyRepository.findOne({
-      email,
-    });
-    const signedUser = await this.userRepository.findUserByEmail(email);
-    if (signedUser) throw new BadRequestException('이미 가입한 유저입니다.');
-    if (currentVerification?.expiredAt) {
-      const { expiredAt, id, token } = currentVerification;
-      if (expiredAt < now) await this.userVerifyRepository.delete(id);
-      if (expiredAt > now) {
-        return {
-          token,
-          email,
-          id,
-        };
-      }
-    }
-    const newVerificationToken = await makeHash(`${email}|${uuidv4()}`);
-    const { token, id } = await this.userVerifyRepository.createOne(email, newVerificationToken);
-
-    return { token, email, id };
-  }
-
-  /**
-   * @returns 이메일 사용자가 이메일을 검증하였는 지 여부를 반환합니다.
-   */
-  public async lastCheckingEmailVerify(email: string) {
-    const alreadyUser = await this.userRepository.findUserByEmail(email);
-    if (alreadyUser) throw new BadRequestException('이미 가입된 이메일입니다.');
-    const verificationInstance = await this.userVerifyRepository.findOne({
-      email,
-    });
-
-    if (!verificationInstance) {
-      return false;
-    }
-
-    return verificationInstance;
-  }
-
-  /**
-   * @return 이메일 형식에 대한 검증 값을 UserVerify Entity 정보로 반환한다.
-   */
-  async verifyEmail(id: string, token: string) {
-    const record = (await this.userVerifyRepository.findOne(id)) as UserVerifyEntity;
-    if (!record) {
-      throw new HttpException('인증 url 이 잘못되었습니다.', HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-    if (record.isVerified) return record;
-    if (record.expiredAt < new Date()) return false;
-
-    const isEqual = token === record.token;
-    if (isEqual) record.isVerified = true;
-    if (!isEqual) record.isVerified = false;
-
-    await record.save();
-    return record;
   }
 
   /**
    * @return 사용자 정보를 입력받아 다양한 예외를 처리하고, 생성정보를 반환합니다.
    */
-  async join({ username, password, email, skills, job }: CreateUserDto) {
-    const verification = await this.userVerifyRepository.findOne({
-      email,
-    });
-    if (!verification || !verification.isVerified) {
-      throw new HttpException('이메일 인증을 수행해주세요.', HttpStatus.UNAUTHORIZED);
-    }
-
-    const currentUser = await this.userRepository.findUserByEmail(email);
-    if (currentUser) {
-      throw new HttpException('이미 존재하는 유저입니다.', HttpStatus.CONFLICT);
-    }
-
-    const currentName = await this.userRepository.findUserByName(username);
-    if (currentName) {
-      throw new HttpException('이미 존재하는 닉네임입니다.', HttpStatus.CONFLICT);
-    }
-    const jobEntity = await this.userJobRepository.fineOneById(job);
-
-    if (jobEntity === undefined) {
-      throw new HttpException('직업 정보가 일치하지 않습니다.', HttpStatus.BAD_REQUEST);
-    }
+  async join(dto: CreateUserDto) {
+    const { username, password, email, skills, job } = dto;
+    await this.authValidateService.validateJoin(dto);
 
     const user = await this.userRepository.createUserOne({
       username,
       password: await makeHash(password),
       email,
       skills,
-      job: jobEntity ? jobEntity.id : null,
+      job: job ? job.id : null,
     } as CreateUserDto);
 
     const { id } = user;
-    const { password: notUsingProps1, hashedRefreshToken: notUsingProps2, ...userProps } = user;
+    const { password: pw, hashedRefreshToken, ...userProps } = user;
     const accessToken = this.getAccessToken({
       id,
       username,
@@ -244,7 +119,7 @@ class AuthService {
   /**
    * @desc RefreshToken 을 사용자 정보에 저장합니다.
    */
-  async saveHashRefreshToken(refreshToken: string, email: string) {
+  async saveHashRefreshToken(refreshToken: string, email: string): Promise<void> {
     const hashedToken = await bcrypt.hash(refreshToken, 12);
     await this.userRepository.update(
       { email },
@@ -257,7 +132,7 @@ class AuthService {
   /**
    * @desc id 에 대응하는 유저정보에 대한 RefreshToken 을 검증합니다.
    */
-  async getUserIfTokenMatches(refreshToken: string, id: number) {
+  async getUserIfTokenMatches(refreshToken: string, id: number): Promise<UserEntity | null> {
     const user = await this.userRepository.findOne(id);
     if (user?.hashedRefreshToken) {
       const isMatch = await bcrypt.compare(refreshToken, user.hashedRefreshToken);
@@ -271,7 +146,7 @@ class AuthService {
   /**
    * @desc JWT AccessToken 에 대한 유효시간을 구해서 현재 시간에서 유효시간이 지난 시간을 반환합니다.
    */
-  getAccessTokenExpirationTime() {
+  getAccessTokenExpirationTime(): Date {
     const accessTokenExpirationTime = this.configService.get(
       'JWT_ACCESS_TOKEN_EXPIRATION_TIME',
     ) as string;
@@ -279,7 +154,7 @@ class AuthService {
     return addMinutes(new Date(), minutes);
   }
 
-  getAccessTokenByHeaders(headers: IncomingHttpHeaders) {
+  getAccessTokenByHeaders(headers: IncomingHttpHeaders): string | undefined {
     return headers.authorization?.split(' ')[1];
   }
 
@@ -287,12 +162,10 @@ class AuthService {
     try {
       const secretKey = this.configService.get('JWT_ACCESS_TOKEN_SECRET') as string;
       const decodedToken = (await jwtVerifyPromise(accessToken, secretKey)) as IJwtPayload;
-      const user = await this.userService.findUserShallow(decodedToken.id);
+      const user = await this.userRepository.findUserShallow(decodedToken.id);
       return { isLoggedIn: true, user };
     } catch {
       return { isLoggedIn: false };
     }
   }
 }
-
-export default AuthService;
